reestructure python files
