--- ml/ml_script/code1.py
@@ -1,1 +1,0 @@
-ALGO = 'DESDE utils'
