+++ ml/ml_script/cvat_smallbox/code.py
@@ -1,35 +1,7 @@
-from xml.etree import ElementTree as ET
 import logging
 import click
 
-
-def get_segments_root(root, v=False):
-    try:
-        segments = []
-        for i in root.iter('segment'):
-            o = {}
-            for j in i:
-                o[f'{j.tag}'] = j.text
-            segments.append(o)
-        if v:
-            click.echo(click.style('--- segments created :) ', fg='bright_cyan', bold=True))
-        return segments
-    except Exception as e:
-        logging.error(e.__str__())
-
-
-def read_xml(in_file, v=False):
-    """
-    open xml file and return root  iterate file
-    """
-
-    try:
-        tree = ET.parse(in_file)
-        if v:
-            click.echo(click.style('--- xml readed :) ', fg='bright_cyan', bold=True))
-        return tree.getroot()
-    except Exception as e:
-        logging.error(e.__str__())
+from ..utils.code import (get_segments_root, read_xml)
 
 
 @click.command()
@@ -81,7 +53,7 @@ def process(in_file, toleranci):
         # df = DataFrame(list_image_err)
         # print version
         for i in list_image_err:
-            print(f'{i[0]},{i[1]},{i[2]}')
+            print(','.join(map(str, i)))
 
 
 if __name__ == '__main__':
